Modify UpgradeOperator fields (#149)
