+++ crates/enclave-contract/src/lib.rs
@@ -6,8 +6,8 @@
 #![cfg_attr(docsrs, feature(doc_cfg, doc_auto_cfg))]
 #![cfg_attr(not(test), warn(unused_crate_dependencies))]
 
-pub mod contracts;
+pub mod contract_interface;
 pub mod deployment;
 
-pub use contracts::*;
+pub use contract_interface::*;
 pub use deployment::*;
